refactor: Import top-level components/ facades from native binary
Move WrapFactory into cmd/jaeger/internal/telemetryfactory/ subpackage
to break the import cycle, allowing components.go to import from the
top-level components/ layer instead of the bridge layer. This ensures
the native binary exercises the full facade chain (components/ →
cmd/jaeger/components/ → cmd/jaeger/internal/).
