+++ cmd/jaeger/components/telemetry/factory.go
@@ -3,13 +3,6 @@
 
 package telemetry
 
-import (
-	coltelemetry "go.opentelemetry.io/collector/service/telemetry"
-	"go.opentelemetry.io/collector/service/telemetry/otelconftelemetry"
+import impl "github.com/jaegertracing/jaeger/cmd/jaeger/internal/telemetryfactory"
 
-	jaegerinternal "github.com/jaegertracing/jaeger/cmd/jaeger/internal"
-)
-
-func NewFactory() coltelemetry.Factory {
-	return jaegerinternal.WrapFactory(otelconftelemetry.NewFactory())
-}
+var NewFactory = impl.NewFactory
